updates to wg contribution
- adds testing code
- adds command line utility code
- adds code to pass round trip test

also there were some interface changes in the update to the new version
of the netlink library that needed to be dealt with. this mainly had to
do with error strings. the ensure and absent functions depend on error
string recognition, and the strings changed in the update so updates to
that code were required.

Signed-off-by: Ryan Goodfellow <rgoodfel@isi.edu>